requirements.txt
bitsandbytes
hf_xet
transformers 
torch 
Pillow 
accelerate
opencv-python
moviepy
ffmpeg
openai-whisper
langchain_google_genai
dotenv
chromadb